backend/requirements.txt
Django==3.1.5
django-filter==2.4.0
django-mptt==0.11.0
django-silk==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
django-admin-list-filter-dropdown==1.0.3
django-ckeditor==6.0.0
djoser==2.1.0
oauthlib==3.1.0
Pillow==8.1.0
PyJWT==2.0.1
django-cors-headers==3.6.0
psycopg2-binary